backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai>=1.40.0
redis==5.0.1
rq==1.15.1
rq-dashboard==0.6.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
numpy==1.26.3
scikit-learn==1.4.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-flask==1.3.0
python-multipart==0.0.6
Werkzeug==3.0.1
pypdf>=3.0.0

